split helloworld
