+++ examples/helloworld/greeter_client/main.go
@@ -19,13 +19,12 @@
 package main
 
 import (
-	"context"
 	"log"
 	"os"
-	"time"
 
 	"google.golang.org/grpc"
 	pb "google.golang.org/grpc/examples/helloworld/helloworld"
+	"google.golang.org/grpc/examples/helloworld/helloworld/client"
 )
 
 const (
@@ -47,11 +46,5 @@ func main() {
 	if len(os.Args) > 1 {
 		name = os.Args[1]
 	}
-	ctx, cancel := context.WithTimeout(context.Background(), time.Second)
-	defer cancel()
-	r, err := c.SayHello(ctx, &pb.HelloRequest{Name: name})
-	if err != nil {
-		log.Fatalf("could not greet: %v", err)
-	}
-	log.Printf("Greeting: %s", r.Message)
+	client.CallSayHello(c, name)
 }
